fix:[用水单位管理]完善修改删除操作

+++ src/main/java/com/zjtc/service/impl/BankServiceImpl.java
@@ -29,6 +29,9 @@ public class BankServiceImpl extends ServiceImpl<BankMapper, Bank> implements
     for (Bank bank : banks) {
       ids.add(bank.getId());
     }
+    if (ids.isEmpty()){
+      return false;
+    }
     return this.baseMapper.deletedBank(ids);
   }
